Moved fonts inside library
